111. Minimum Depth of Binary Tree
