+++ src/main/kotlin/me/bossm0n5t3r/leetcode/minimumDepthOfBinaryTree/MinimumDepthOfBinaryTree.kt
@@ -1,0 +1,26 @@
+package me.bossm0n5t3r.leetcode.minimumDepthOfBinaryTree
+
+import me.bossm0n5t3r.leetcode.utils.TreeNode
+import java.util.LinkedList
+
+class MinimumDepthOfBinaryTree {
+    class Solution {
+        fun minDepth(root: TreeNode?): Int {
+            if (root == null) return 0
+            var minimumDepth = 0
+            val queue = LinkedList<TreeNode?>()
+            queue.add(root)
+            while (queue.isNotEmpty()) {
+                val size = queue.size
+                minimumDepth++
+                repeat(size) {
+                    val cur = queue.poll()
+                    if (cur?.left == null && cur?.right == null) return minimumDepth
+                    cur.left?.let { queue.add(it) }
+                    cur.right?.let { queue.add(it) }
+                }
+            }
+            return minimumDepth
+        }
+    }
+}
